requirements.txt
numpy==1.26.2
opencv-python==4.9.0.80
pandas==2.1.4
Pillow==10.0.1
scipy==1.11.4
tensorboard==2.12.1
torch==2.5.1
torchaudio==2.5.1
torchinfo==1.8.0
torchvision==0.20.1
tqdm==4.65.0
packaging==24.0